+++ hw07-design-patterns/src/main/java/ru/otus/hw07/services/ATMImpl.java
@@ -4,7 +4,6 @@ import ru.otus.hw07.domain.Banknote;
 import ru.otus.hw07.domain.NoSuitableBanknotesAvailableException;
 import ru.otus.hw07.domain.NotSufficientFundsException;
 
-import java.math.BigInteger;
 import java.util.Collections;
 import java.util.List;
 
@@ -31,7 +30,7 @@ public class ATMImpl implements ATM {
     }
 
     @Override
-    public BigInteger getAvailableMoneyCount() {
+    public long getAvailableMoneyCount() {
         return moneyStorageService.getAvailableMoneyCount();
     }
 }
