+++ hw07-design-patterns/src/main/java/ru/otus/hw07/services/MoneyStorageService.java
@@ -5,12 +5,11 @@ import ru.otus.hw07.domain.Denomination;
 import ru.otus.hw07.domain.NoSuitableBanknotesAvailableException;
 import ru.otus.hw07.domain.NotSufficientFundsException;
 
-import java.math.BigInteger;
 import java.util.List;
 
 public interface MoneyStorageService {
     long storeMoney(List<Banknote> moneyBundle);
     List<Banknote> retrieveMoney(long amount) throws NoSuitableBanknotesAvailableException, NotSufficientFundsException;
-    BigInteger getAvailableMoneyCount();
+    long getAvailableMoneyCount();
     int getAvailableBanknotesCount(Denomination banknoteDenomination);
 }
